+++ md_utils/fes_combo.py
@@ -1,7 +1,7 @@
 from __future__ import print_function
 import logging
 
-from raptor_utils.common import find_files_by_dir
+from md_utils.common import find_files_by_dir
 
 __author__ = 'cmayes'