+++ tests/test_fes_combo.py
@@ -2,8 +2,8 @@ import logging
 import os
 import unittest
 
-from raptor_utils.common import find_files_by_dir
-from raptor_utils.fes_combo import combine, DEF_FILE_PAT, DEF_TGT, map_fes, extract_header, write_combo
+from md_utils.common import find_files_by_dir
+from md_utils.fes_combo import combine, DEF_FILE_PAT, DEF_TGT, map_fes, extract_header, write_combo
 
 __author__ = 'cmayes'