+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 -------
 
 0.1.0 (2015-09-01)
----------------------
+------------------
 
-* First release on PyPI.
+* Renamed to md_utils
+* Added fes_combo.py